src/requirements.txt
alembic==1.10.2
asyncpg==0.27.0
Faker==18.3.1
fastapi==0.94.1
httpx==0.23.3
orjson==3.8.7
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.6
pytest==7.2.2
pytest_asyncio==0.21.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.7
SQLAlchemy-Utils==0.40.0
uvicorn[standard]==0.21.0
